Add optimize_amplitude
Port the functionality from matlab.  This function will have to be
refactored once we figure out how to integrate the calibration
loop into auspex.  For now, the "loop" is simulated using the
update_data function pointer passed into the routine.  This simulates
receving a sequence from the digitizer with an updated pulse
amplitude for testing.