Make Bokeh server process persist after quitting.
Writes the Bokeh server PID to file upon loading the server. When launching a
new experiment, we check for the existence of a running process on that PID, and
skip loading the Bokeh server if we find it.

An open issue is to figure out how/when to shut the Bokeh server down. Ideally,
it would timeout if it hadn't received any messages after XX seconds. Not sure
how to make that happen, though.

Also, this code does not "double fork" the Bokeh process. Consequently, if the
user sends SIGINT with Ctrl-C, that will shut down the Bokeh server. Not sure if
you want to think of that as a feature or a bug.