Alazar: unregister sockets on disconnect
The problem occurred when *changing* which channel was being used
without unloading/reloading the libalazar library. This caused the
library to have a dangling socket reference that it would try to stuff
data into, but there would be no one listening on the other end. We
added a means to unregister all sockets to clean up properly.

--with @grahamrow