removed some debugging code; tightened up waiting
if the total observed points taken by the output channels
sums to less than the total fake data injected (which the
X6 object knows, because it injected it), then wait for
a short period of time and then check again, ad nauseum.

Note that we assume that the points will eventually be taken:
there's no other way to break out of this loop except by
seeing all the expected data.  If we don't see it all, then
we go into an infinite loop.